Various rearrangements and fixes.
I've gone down the silly route of allowing a flavour to be applied after
the customisation of the VM for VCloud, when actually I should do the
customisation first then have a generic VCloud importer.  But anyway
this works for Docker just now and is very easily ripped out later.